:truck: refactor: move controllers to directory
