+++ src/domain/user/user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
-import { UserController } from './user.controller';
-import { MysqlModule } from '../../database/mysql.module';
+import { UserController } from '../../controllers/user.controller';
+import { MysqlModule } from '../../infra/database/mysql.module';
 
 @Module({
   imports: [MysqlModule],
